fix: resolve AArch64 mmap cast and test cleanup errors
- Remove unnecessary offset as i64 cast in raw_mmap (line 917)
- Fix repro_shim_init_race.sh cleanup to ignore permission errors
- All Docker CI tests now pass

+++ crates/vrift-shim/src/syscalls/linux_raw.rs
@@ -914,7 +914,7 @@ pub unsafe fn raw_mmap(
             in("x2") prot as i64,
             in("x3") flags as i64,
             in("x4") fd as i64,
-            in("x5") offset as i64,
+            in("x5") offset,
             lateout("x0") ret,
         );
         ret as *mut c_void
